refactor(core): Consolidate handlers and controllers
