+++ src/main/java/org/sustain/controllers/OsmController.java
@@ -1,4 +1,4 @@
-package org.sustain.datasets.controllers;
+package org.sustain.controllers;
 
 import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoCollection;
@@ -19,7 +19,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.LinkedBlockingQueue;
 
-import static org.sustain.datasets.controllers.SpatialQueryUtil.getGeometryFromGeoJson;
+import static org.sustain.controllers.SpatialQueryUtil.getGeometryFromGeoJson;
 
 public class OsmController {
     private static final Logger log = LogManager.getLogger(OsmController.class);
