+++ src/main/java/org/sustain/controllers/SpatialQueryUtil.java
@@ -1,4 +1,4 @@
-package org.sustain.datasets.controllers;
+package org.sustain.controllers;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonArray;
