+++ src/main/java/org/sustain/handlers/RegressionQueryHandler.java
@@ -1,9 +1,10 @@
-package org.sustain.modeling;
+package org.sustain.handlers;
 
 import io.grpc.stub.StreamObserver;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.sustain.*;
+import org.sustain.modeling.SustainLinearRegression;
 import org.sustain.util.Constants;
 
 public class RegressionQueryHandler {
