+++ src/main/java/org/sustain/querier/Querier.java
@@ -1,29 +1,20 @@
 package org.sustain.querier;
 
-import com.mongodb.client.FindIterable;
 import com.mongodb.client.AggregateIterable;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoCursor;
 import com.mongodb.client.MongoDatabase;
-import com.mongodb.client.model.Filters;
 import com.mongodb.BasicDBObject;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import org.bson.Document;
-import org.bson.conversions.Bson;
 
 import org.json.JSONArray;
-import org.json.JSONObject;
-import com.google.gson.Gson;
 
-import org.sustain.JoinOperator;
-import org.sustain.ComparisonOperator;
-import org.sustain.CompoundResponse;
-import org.sustain.CompoundRequest;
 import org.sustain.Query;
-import org.sustain.util.Constants;
+import org.sustain.handlers.CompoundQueryHandler;
 
 import org.sustain.db.mongodb.DBConnection;
